requirements.txt
Flask  
datetime 
uuid
PyJWT
hashlib
os